EOS
Added most of the add command.
Now I gotta figure out how the item/data is stored in the "customitemsconfig.yml" cause it seems that it is not automatically added to it.
In addition the adding needs some polishing, surely some bug fixes and generally needs to be more responsive. Adding some sounds/simple info should help alot.