EOS
Working on a Filter for logging the error message that appears, due to "bannable" not being a valid data tag in the `config.yml` file of the BanItems plugin.
I want to remove it both from the console and the ingame chat.
Therefore I should use a form of the Strategy structure, to have the same algorithm be applicable both to ingame and console.