Fixed issue #1
While still not perfect, it is an improvement.
The formatting of the console message is now correct, there is the possibility that the message is not in the correct order.
Aka, the "[BanItem] Successfully loaded ..." message might have already been sent.
This is sadly due to the asynchronous nature of the logger and I don't want to throttle it, sooooo, live with a bit of imperfection.